v3

+++ ft_memset.c
@@ -6,21 +6,21 @@
 /*   By: chevrethis <chevrethis@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/29 13:15:16 by chevrethis        #+#    #+#             */
-/*   Updated: 2025/04/29 13:15:18 by chevrethis       ###   ########.fr       */
+/*   Updated: 2025/04/29 14:26:47 by chevrethis       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void    *ft_memset(void *s, int c, size_t n)
+void	*ft_memset(void *s, int c, size_t n)
 {
-    int i;
+    size_t	i;
 
-    i = 0;
-    while (i < n)
-    {
-        *(unsigned char*)(s + i) = (unsigned char)c;
-        i++;
-    }
-    return (s);
+	i = 0;
+	while (i < n)
+	{
+		*(unsigned char*)(s + i) = (unsigned char)c;
+		i++;
+	}
+	return (s);
 }
